Introduce explicit `GetFields` accessor
Tinkering as part of a reflection on lukasschwab.me/blog/gen/elixir-logger-magic.

+ [break] Make `key` a private type: private singleton implementation.
+ ~[bug] `Append` new fields over existing: prefer later fields.~ This
  is not a bug: `GetField` (as defined upstream) prefers the first entry.
+ Add `GetFields` for getting several k/v pairs with explicit
  expectations. Marshal unfound expected keys as a zax-managed field.
+ Add `Prune` for eliminating overwritten fields: O(n) for the number of
  fields.
+ [break] Return an `ok` boolean from GetField to indicate found/not found.
+ [break] Renames `Get` to `GetAll`.

As written, this is a breaking change.

Idea: maybe this should simulate a stack per key, so that you can
push/pop values? What's the right way to serialize that: whole stack, or
just the head?

I considered adding a substitute `Key` type (instead of string) required
for setting key/value pairs here, but that complicates interoperability
with zap. Could consider a thin wrapper around Zap with a different
`key` type, but it's such a complex interface. In theory, string
conversion only needs to happen at marshall...

In any case, one could enforce this in a linter: no string literals in
zap field keys. That's a useful principle (makes it easier to track
which keys are shared and where they're set), but it's not worth
stressing here.

Ultimately, this fork doesn't achieve much besides the bug fix and the
check for key existence in the context. I think *that* is an important
part of using the ctx to pass values around.